requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
requests
google-genai